fixed data type problem

+++ buildPyramid.m
@@ -5,7 +5,7 @@ function [ levels, sizes ] = buildPyramid( image  )
     kernel = kernel/ sum(sum(kernel));
 
 
-    currentG = image;
+    currentG = int16(image);
     % filter and downsample
     nextG =  imresize(imfilter(currentG, kernel, 'replicate'),0.5,'nearest');
     % upsample and subtract
